split logic from LegacyRepository into method, move to parent class
split logic from LegacyRepository._get_release_info for getting sha256 hash from file when link hash is not sha* into its own method and move it to parent class PyPiRepository for reuse
